requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
sqlite3

# Vector database and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1
python-magic==0.4.27

# Text processing and NLP
nltk==3.8.1
spacy==3.7.2
transformers==4.36.0
torch==2.1.0

# Security and guardrails
openai==1.3.0
tiktoken==0.5.1
regex==2023.10.3

# Evaluation and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pyyaml==6.0.1
scikit-learn==1.3.2

# Web UI
streamlit==1.28.1
gradio==4.7.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
